Move ttest functions to stat (from scil_compare_connectivity)
