+++ scilpy/stats/stats.py
@@ -3,6 +3,7 @@
 import logging
 
 from itertools import combinations
+
 import scipy.stats
